v2_SAM+NN/requirements.txt
transformers
pillow
matplotlib
numpy
opencv-python
tqdm
wandb